correct a bug in makefiles concerning fat wheels
The library _morfeusz2.so was reused in two contexts, only one of
which requiring an RPATH setting.  Thus, the result depended on the
order in which make targets were specified.  Now the RPATH compilation
is performed in a separate directory (fatlib) and the makefile “knows”
it has to be compiled separately.